LayoutTests/editing/selection/context-menu-text-selection-lookup-expected.txt

CONSOLE MESSAGE: line 31: ReferenceError: Can't find variable: expectedText
This test checks that triggering the context menu selects/doesn't select using surround text as context (as platform-appropriate).

New York
RUNNING
RUNNING
RUNNING
